requirements.txt
transformers>=4.35.0
langchain>=0.1.0
langchain-community>=0.0.35
streamlit
chromadb>=0.4.0
sentence-transformers>=2.2.2
torch>=2.1.0
python-dotenv>=1.0.0
huggingface-hub>=1.25.1
python-dotenv
langchain-huggingface